modifying parse structure --no bonus
